explore-assistant-cloud-function/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-bigquery==3.15.0
looker-sdk>=23.0.0
PyJWT==2.8.0
pandas>=1.3.0
json5>=0.9.0  # Flexible JSON parsing for LLM outputs
pydantic>=1.10.0  # Data validation and parsing for robust schema handling
demjson3>=3.0.0  # Robust JSON decoding for malformed LLM outputs

# MCP (Model Context Protocol) dependencies
mcp>=0.5.0  # Core MCP library
